pyproject.toml
[tool.poetry]
name = "kubecontext-merger"
version = "0.1.0"
description = ""
authors = ["Nikita <nikitosnikn@yandex.ru>"]
license = "MIT"
readme = "README.md"
packages = [{include = "kubecontext_merger"}]

[tool.poetry.dependencies]
python = "^3.11"
typer = "^0.7.0"
PyYAML = "^6.0"

[tool.poetry.scripts]
merger = "kubecontext_merger:app"



[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
